+++ lib/http/service.js
@@ -66,7 +66,6 @@ module.exports = (container) => {
   require('../resources/comments')(service, endpoint);
   require('../resources/analytics')(service, endpoint);
   require('../resources/datasets')(service, endpoint);
-  require('../resources/datasets-temp')(service, endpoint);
 
   ////////////////////////////////////////////////////////////////////////////////
   // POSTRESOURCE HANDLERS
