--- lib/resources/datasets-temp.js
@@ -1,35 +1,0 @@
-// Copyright 2022 ODK Central Developers
-// See the NOTICE file at the top-level directory of this distribution and at
-// https://github.com/getodk/central-backend/blob/master/NOTICE.
-// This file is part of ODK Central. It is subject to the license terms in
-// the LICENSE file found in the top-level directory of this distribution and at
-// https://www.apache.org/licenses/LICENSE-2.0. No part of ODK Central,
-// including this file, may be copied, modified, propagated, or distributed
-// except according to the terms contained in the LICENSE file.
-
-const sanitize = require('sanitize-filename');
-
-const { streamEntityCsvs } = require('../data/entity');
-const { contentDisposition } = require('../util/http');
-const { getOrNotFound } = require('../util/promise');
-
-module.exports = (service, endpoint) => {
-
-  //////////////////////////////////////////////////////////////////////////////
-  // Dataset resources
-  // Temporary so as not to conflict with other resources developed at the same time
-
-  service.get('/projects/:projectId/datasets/:name/download', endpoint(({ Datasets, Entities, Projects }, { params, auth }, _, response) =>
-    Projects.getById(params.projectId) // TODO: different lookup/auth scheme for datasets
-      .then(getOrNotFound)
-      .then((project) => auth.canOrReject('project.read', project))
-      .then(() => Datasets.getByName(params.name, params.projectId)
-        .then((dataset) => Entities.streamForExport(dataset.id)
-          .then((entities) => {
-            const filename = sanitize(dataset.name);
-            const extension = 'csv';
-            response.append('Content-Disposition', contentDisposition(`${filename}.${extension}`));
-            response.append('Content-Type', 'text/csv');
-            return streamEntityCsvs(entities, dataset.properties);
-          })))));
-};
