+++ lib/resources/datasets.js
@@ -6,9 +6,13 @@
 // https://www.apache.org/licenses/LICENSE-2.0. No part of ODK Central,
 // including this file, may be copied, modified, propagated, or distributed
 // except according to the terms contained in the LICENSE file.
+
+const sanitize = require('sanitize-filename');
 const { getOrNotFound } = require('../util/promise');
 const { Form } = require('../model/frames');
 const { ensureFormDef } = require('../util/db');
+const { streamEntityCsvs } = require('../data/entity');
+const { contentDisposition } = require('../util/http');
 
 module.exports = (service, endpoint) => {
   service.get('/projects/:id/datasets', endpoint(({ Projects, Datasets }, { auth, params, queryOptions }) =>
@@ -23,4 +27,18 @@ module.exports = (service, endpoint) => {
       .then((form) => auth.canOrReject('form.update', form)) // what verb is required here? a new one
       .then(ensureFormDef)
       .then(() => Datasets.getDatasetDiff(params.id))));
+
+  service.get('/projects/:projectId/datasets/:name/download', endpoint(({ Datasets, Entities, Projects }, { params, auth }, _, response) =>
+    Projects.getById(params.projectId) // TODO: different lookup/auth scheme for datasets
+      .then(getOrNotFound)
+      .then((project) => auth.canOrReject('project.read', project))
+      .then(() => Datasets.getByName(params.name, params.projectId)
+        .then((dataset) => Entities.streamForExport(dataset.id)
+          .then((entities) => {
+            const filename = sanitize(dataset.name);
+            const extension = 'csv';
+            response.append('Content-Disposition', contentDisposition(`${filename}.${extension}`));
+            response.append('Content-Type', 'text/csv');
+            return streamEntityCsvs(entities, dataset.properties);
+          })))));
 };
